requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose
langchain
langgraph
langchain-openai